Replace Contao fragment renderer with service decorating the Symfony fragment handler
